tidying: rm .DS_store files; rm executable flags
Having system-specific binary files ".DS_store" is not wise
in a source distribution (and, in principle, under the GPL,
the sources of those binary files should be available - since they
presumably have some sort of info on directory structure/history,
it could be argued that the info is already available).

Giving executable status to text files such as *.py files
doesn't seem to be wise either. These are text files that
are to be interpreted by a python binary, not binary files
to be run directly. So this commit removes executable flags
from some python files.