fix bug of accessing the array orig at invalid indices
In src/voz1b1.c the array orig can be accessed at an invalid index,
yielding what to the user appears to be a programming error but in
reality is an unwise choice of analysis parameters.

This commit solves the bug by initialising a boolean array
`orig_exists` to have all members false. This array is updated along
with the two lines that evaluate `orig`. When `orig` is used later
on, `orig_exists` is first checked to see if this is valid. If it's
invalid, the user is warned and a fatal error occurs.

(This commit also tidies up some trailing whitespace.)